frontend/hub: fix fallback customization values and expose the llm markup for the price estimate; refactoring
